Added Snapshot Menu
User can now change snapshots by holding down the middle
foot switch on the main menu.  This also saves the current
fx that were turned on or off in the different snapshots.  This
information is retained until the user switches presets again.

Also, added a cap on how many hipbox midi notes could be
sent.  The max preset that midi data can be sent on is 41.  This
is because on preset 41, midi notes 123,124,125 are sent and
preset 42 would have to use 126,127,128 and note 128 and greater
do not exist.