Refactored code to add in better button debounce
I was having a severe issue with buttons auto-triggering multiple
times after a single press.  After separating the ground
wire of the buttons from the LED's, it improved some but I still
needed software means to remove it.

What I ultimately did was added a 100ms debounce to the button
after it registered a button press as well as before.  When having
this sandwich of debounce, it seems to be running much more
smoothly.