Changed preset menu back to the old way
My first intentions were to keep all the menus similar in
the way they operate, switch 3 long-press would be the back
button essentially.  Well in practice, that wasn't very
intuitive when trying to switch presets quickly, so I switched
it back to the original way where switch 1 and 3 are up and down,
middle short press is go back to main menu, and long press on
middle will go to preset 0.